+++ nion/utils/ThreadPool.py
@@ -5,11 +5,13 @@ Useful classes for handling threads.
 # standard libraries
 import concurrent.futures
 import copy
+import dataclasses
 import logging
 import queue
 import threading
 import time
 import typing
+import weakref
 
 # third party libraries
 # None
@@ -187,55 +189,63 @@ class ThreadPool(object):
             self.__queue.task_done()
 
 
+@dataclasses.dataclass
+class DispatcherInfo:
+    is_dispatching_lock: threading.RLock
+    is_dispatch_pending: bool
+    dispatch_future: typing.Optional[concurrent.futures.Future]
+    dispatch_thread_cancel: threading.Event
+    cached_value_time: float
+
+
 class SingleItemDispatcher:
     """Dispatch a function to the thread pool, ensuring only one is running at once."""
-
     def __init__(self, *, executor: typing.Optional[concurrent.futures.ThreadPoolExecutor] = None, minimum_period: float = 0.0):
         self.__executor = executor or concurrent.futures.ThreadPoolExecutor()
         self.__minimum_period = minimum_period
-        self.__is_dispatching_lock = threading.RLock()
-        self.__is_dispatch_pending = False
-        self.__dispatch_future: typing.Optional[concurrent.futures.Future] = None
-        self.__dispatch_thread_cancel = threading.Event()
-        self.__cached_value_time = 0
-        self.on_computed: typing.Optional[typing.Callable[[], None]] = None
+        self.__dispatcher_info = DispatcherInfo(threading.RLock(), False, None, threading.Event(), 0.0)
+
+        def finalize(dispatcher_info: DispatcherInfo, s: str) -> None:
+            recompute_future = dispatcher_info.dispatch_future  # avoid race by using local
+            if recompute_future:
+                dispatcher_info.dispatch_thread_cancel.set()
+                concurrent.futures.wait([recompute_future])
+
+        weakref.finalize(self, finalize, self.__dispatcher_info, str(self))
 
     def close(self) -> None:
-        self.on_computed = None
-        recompute_future = self.__dispatch_future  # avoid race by using local
-        if recompute_future:
-            self.__dispatch_thread_cancel.set()
-            concurrent.futures.wait([recompute_future])
-
-    def __dispatch_task(self, fn: typing.Callable[[], None]) -> None:
-        while True:
-            try:
-                if self.__dispatch_thread_cancel.wait(0.05):  # gather changes and helps tests run faster
-                    return
-                minimum_time = self.__minimum_period
-                current_time = time.time()
-                if current_time < self.__cached_value_time + minimum_time:
-                    if self.__dispatch_thread_cancel.wait(self.__cached_value_time + minimum_time - current_time):
-                        return
-                self.__is_dispatch_pending = False  # any pending calls up to this point will be realized in the recompute
-                fn()
-            finally:
-                with self.__is_dispatching_lock:
-                    # the only way the thread can end is if not pending within lock.
-                    # recompute_future can only be set within lock.
-                    if not self.__is_dispatch_pending:
-                        self.__dispatch_future = None
-                        break
+        pass
 
     def dispatch(self, fn: typing.Callable[[], None]) -> concurrent.futures.Future:
         # dispatch the function on a thread.
         # if already executing, ensure the thread dispatch again.
         # may be called on the main thread or a thread - must return quickly in both cases.
-        with self.__is_dispatching_lock:
+        with self.__dispatcher_info.is_dispatching_lock:
             # in case thread is already running, set pending.
             # the only way the thread can end is if not pending within lock.
             # dispatch_future can only be set within lock.
-            self.__is_dispatch_pending = True
-            if not self.__dispatch_future:
-                self.__dispatch_future = self.__executor.submit(self.__dispatch_task, fn)
-            return self.__dispatch_future
+            self.__dispatcher_info.is_dispatch_pending = True
+            if not self.__dispatcher_info.dispatch_future:
+
+                def dispatch_task(fn: typing.Callable[[], None], minimum_time: float, dispatcher_info: DispatcherInfo) -> None:
+                    while True:
+                        try:
+                            if dispatcher_info.dispatch_thread_cancel.wait(0.05):  # gather changes and helps tests run faster
+                                return
+                            current_time = time.time()
+                            if current_time < dispatcher_info.cached_value_time + minimum_time:
+                                if dispatcher_info.dispatch_thread_cancel.wait(dispatcher_info.cached_value_time + minimum_time - current_time):
+                                    return
+                            dispatcher_info.is_dispatch_pending = False  # any pending calls up to this point will be realized in the recompute
+                            fn()
+                            dispatcher_info.cached_value_time = time.time()
+                        finally:
+                            with dispatcher_info.is_dispatching_lock:
+                                # the only way the thread can end is if not pending within lock.
+                                # recompute_future can only be set within lock.
+                                if not dispatcher_info.is_dispatch_pending:
+                                    dispatcher_info.dispatch_future = None
+                                    break
+
+                self.__dispatcher_info.dispatch_future = self.__executor.submit(dispatch_task, fn, self.__minimum_period, self.__dispatcher_info)
+            return self.__dispatcher_info.dispatch_future