+++ CHANGES.rst
@@ -3,6 +3,7 @@ Changelog (nionutils)
 
 0.3.27 (UNRELEASED)
 -------------------
+- Remove unused/deprecated PersistentObject.
 - Add property changed property model for monitoring observables.
 - Change Recorder to allow customer logger.
 - Change ThreadPool to be no-close.